+++ rrule/src/core/rruleset.rs
@@ -1,5 +1,5 @@
 use super::{datetime::DateTime, rrule::RRule};
-use crate::{parser::build_rruleset, RRuleError};
+use crate::{parser::build_rruleset, RRuleError, WithError};
 use chrono::TimeZone;
 use chrono_tz::UTC;
 use std::str::FromStr;
@@ -49,32 +49,30 @@ impl RRuleSet {
         self.into_iter().take(limit as usize).collect()
     }
 
-    // /// TODO: **Work in progress**
-    // /// Returns all the recurrences of the rrule.
-    // /// Limit must be set in order to prevent infinite loops.
-    // /// The max limit is `65535`. If you need more please use `into_iter` directly.
-    // ///
-    // /// In case where the iterator ended with an errors the error will be included,
-    // /// otherwise the second value of the return tuple will be `None`.
-    // pub fn all_with_error(&self, _limit: u16) -> (Vec<DateTime>, Option<RRuleError>) {
-    //     let mut iterator = self.into_iter();
-    //     let mut list = vec![];
-    //     let err = None;
-    //     for _i in 0..limit {
-    //         let next = iterator.next();
-    //         match next {
-    //             Some(value) => list.push(value),
-    //             None => {
-    //                 // TODO add error handling in RRuleSetIter
-    //                 // if iterator.has_err() {
-    //                 //     err = iterator.get_err().clone();
-    //                 // }
-    //                 break;
-    //             }
-    //         }
-    //     }
-    //     (list, err)
-    // }
+    /// Returns all the recurrences of the rrule.
+    /// Limit must be set in order to prevent infinite loops.
+    /// The max limit is `65535`. If you need more please use `into_iter` directly.
+    ///
+    /// In case where the iterator ended with an errors the error will be included,
+    /// otherwise the second value of the return tuple will be `None`.
+    pub fn all_with_error(&self, limit: u16) -> (Vec<DateTime>, Option<RRuleError>) {
+        let mut iterator = self.into_iter();
+        let mut list = vec![];
+        let mut err = None;
+        for _i in 0..limit {
+            let next = iterator.next();
+            match next {
+                Some(value) => list.push(value),
+                None => {
+                    if iterator.has_err() {
+                        err = iterator.get_err();
+                    }
+                    break;
+                }
+            }
+        }
+        (list, err.cloned())
+    }
 
     /// Returns the last recurrence before the given datetime instance.
     /// The inc keyword defines what happens if dt is an recurrence.